requirements.txt
annotated-types==0.7.0
anyio==4.9.0
av==14.4.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.6.0
dotenv==0.9.9
faster-whisper==1.1.1
ffmpeg-python==0.2.0
filelock==3.18.0
Flask==3.1.1
flatbuffers==25.2.10
fsspec==2025.5.1
future==1.0.0
google-auth==2.40.2
google-genai==1.16.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.0
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
natsort==8.4.0
numpy==2.2.6
onnxruntime==1.22.0
opencv-python==4.11.0.86
packaging==25.0
platformdirs==4.3.8
protobuf==6.31.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
pyee==13.0.0
pyreadline3==3.5.4
python-dotenv==1.1.0
pytubefix==9.0.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
scenedetect==0.6.6
setuptools==80.8.0
sniffio==1.3.1
sympy==1.14.0
tokenizers==0.21.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
websockets==15.0.1
Werkzeug==3.1.3
